requirements.txt
streamlit==1.35.0
pandas==2.2.2
google-cloud-vision==3.5.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
pyspellchecker==0.7.0
Pillow==10.3.0
